+++ lib/src/value_objects/email_address.dart
@@ -9,11 +9,9 @@ class EmailAddress extends ValueObject<String> {
 
   const EmailAddress._(this.value);
 
-  factory EmailAddress(final String input) {
-    return EmailAddress._(
-      Validator.string()
-          .matches(r"""^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$""")
-          .validate(input),
-    );
-  }
+  factory EmailAddress(final String input) => EmailAddress._(
+    Validator.string()
+        .matches(r"""^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$""")
+        .validate(input),
+  );
 }