+++ lib/src/value_objects/unique_id.dart
@@ -1,11 +1,9 @@
-import 'package:dart_foundation/src/failures/value_failure.dart';
-import 'package:dart_foundation/src/types/value_object.dart';
+import 'package:dart_foundation/dart_foundation.dart';
 import 'package:dartz/dartz.dart';
 import 'package:uuid/uuid.dart';
 
 /// UniqueId [ValueObject]
 class UniqueId extends ValueObject<String> {
-
   /// The value representing Either a ValueFailure or the UniqueId
   @override
   final Either<ValueFailure, String> value;
@@ -15,15 +13,13 @@ class UniqueId extends ValueObject<String> {
 
   /// Factory method for applying a new UniqueId.
   /// This method is calling the UUID service to generate a new UniqueId.
-  factory UniqueId() {
-    return UniqueId._(right(const Uuid().v4()));
-  }
+  factory UniqueId.uuidV4() =>
+      UniqueId._(Validator.unchecked(const Uuid().v4()));
 
   /// Returns the input string as UniqueId.
   ///
   /// There is no possible way to check if an id is unique or not.
   /// Therefore we have to trust that we only use it if the know it is.
-  factory UniqueId.fromUniqueString(final String uniqueId) {
-    return UniqueId._(right(uniqueId));
-  }
+  factory UniqueId.fromUniqueString(final String uniqueId) =>
+      UniqueId._(Validator.unchecked(uniqueId));
 }