+++ CHANGELOG.md
@@ -3,6 +3,12 @@ ___
 
 Copyright: Christian Holler
 
+## 1.0.0+1
+- Added *unchecked()* method to validator, so values can be passed without checks if needed.
+- Added IntValidator class.
+- Added BoolValidator class.
+- Removed factory name '*apply*' from all value objects, failures and any other class. Naming the 
+factory is now only needed if it makes sense.
 
 ## 1.0.0
 - Initial version.